requirements.txt
numpy>=1.20.0
matplotlib>=3.3.0
pandas>=1.2.0
scipy>=1.6.0
pyyaml>=5.4.0  # Optional: for YAML configuration support
